Update Convert-SPWebApplication.md
Added second example for a post-patch required parameter. Filled in missing descriptions for switches. Added valid values for the -To and -From switches. Moved the -From parameter to be more prominent in the list of parameters as it is required. Added to the description to answer a common question after claims conversions.