Update Get-SPOAppInfo.md
Added an extra example and changed the heading for parameters to required. On docs.microsoft.com the name and productid parameters where descriped as optional instead of required.

I didn't noticed these where required until I saw this text in the description: 
Either ProductId or Name must be given. Name is ignored if ProductId is specified.