Update Set-TeamsRetentionCompliancePolicy.md
Made a few changes to the example and clarified TeamsChannelLocation at a few places in the article. Chris, the TeamsChannelLocation literally represents a Group Mailbox (or team mailbox) and what it means is this location targets all the teams channel messages that got ingested into that mailbox. A Team can have several channels and all of their messages get ingested into this one group mailbox. So literally this location param is a group mailbox but we had to use a slightly different name and align with the UI - so we called it TeamsChannelLocation