Update Set-DynamicDistributionGroup.md
This is via content request 80273: 
The documentation doesn't specify that using a wildcard in the middle of the string will resolve it as a literal. However wildcards can be used in the middle of strings like in ad user. To make this documentation more complete, we should specify exactly how these are to be used, and no only how they can't be used in the cloud. 
I worked on a case for weeks with a customer where we went back and forth on the documentation, as it doesn't say anywhere that wildcards can't be used in the middle of the string, but they can be in other products. To show this as proof to those attempting to do this in the future, this update would be helpful.