CloudRecordingServiceSupport cannot set with Set-CsConferencingPolicy
I requested content update when this page was on public in TechNet, Content Idea Request 64449.
When the content moves to doc.microsoft.com and updated part seems reverted. 

Based on following bug, it seems the parameter, -CloudRecordingServiceSupport has been removed from Get-CsConferencingPolicy and no longer works.
  Bug: Office15 3491382 - CloudRecordingServiceSupport parameter in LYO Conference Policy

When we set -CloudRecordingServiceSupport parameter with  Set-CsConferencingPolicy command, we don't get any error but cannot get the value of this parameter from running Get-CsConferencingPolicy.