Suggested Updates to New-PnPTenantSite.md
The LCID references do not list the ID codes for creating a site with a specific LocaleID with a type of UINT32 so have added a command to get that from SharePoint online. Minor corrections to the other commands and seemingly corrected a bad URL at the bottom of the article.