Update Set-DistributionGroup.md
When validating the value passed to the 'ExpansionServer' attribute, the ValidateExpansionServer method is evaluated.
This method takes a single Exchange server only (not an array), so how can a high availability solution be deployed here?
Is this part of older documentation that got carried over? From 2013 onward, our recommendation is to allow any Mailbox server to expand.