Update Set-CsCallQueue.md
@kenwith cc: @roykuntz 

Updating the cmdlet sample for setting music on hold to use stored variable for the audio file ID. This is the only way to do this because there is no Get cmdlet to retrieve stored files. Also added reference to the Import-CsOnlineAudioFile cmdlet which shows using a stored variable when creating but not why or how to use it. Will add a note on that page as well to cross reference back to here.