Update filters-v2.md
Like and Notlike operators have some difference as they are not fully supported in all scenarios i.e. when multiple wildcards are used, they are not supported. As it is mentioned below the operator list.

Please take a call if you would like to highlight those 2 operators differently in the list.