Update Get-CsUserPolicyAssignment.md
Changed the location of the changed information to the -Identity section. Removed examples for Azure AD module and simplified the information provided. I have still included a small comment - "(taken from Azure AD)" - as when my engineers were trying to do this in a hybrid deployment envrionment, they were attempting the use the on premise ObjectGUID and not the Azure AD ObjectID.