Update exchange-online-powershell-v2.md
@JeremyTBradshaw , I agree with what you did, but I did it in a different way.

If you install the module from an elevated PowerShell window vs. `-Scope CurrentUser`, does that then dictate how you need to update the module in the future using Update-Module? I called that out on lines 240 and 254. I was hoping there was some way to easily tell which method was used to originally install the module, but I couldn't find anything.

Feel free to edit my edits.