Update Grant-CsApplicationAccessPolicy.md
We can only assign 1 application access policy at a time for a user. Assigning the new application access policy will override the existing application access policy if any. Just wanted to capture this point as part of note so that anyone don't override the existing assigned policy with this command.