Update Set-CsTeamsCallingPolicy.md
We are in the process of integrating two new settings under TeamsCallingPolicy. While the policy settings are available via the Powershell cmdlets, these settings are not yet ready to be consumed via Teams customers. Adding "reserved for internal Microsoft use" to avoid any confusion related to public use of these two settings.