Update New-Mailbox.md
In the EnableRoomMailboxAccount parameter set, I believe there was an extra opening square bracket '[' making it ambiguous whether `MicrosoftOnlineServiceID` was a required parameter or not.  In all other parameter sets where the parameter appears, it is a named, required parameter.  
Also, a number of parameters in the same parameter set did not have double trailing spaces which can lead to unexpected rendering.