Update Add-RoleGroupMember.md
Confirmed with the engineering team who manages the UI in the Microsoft Purview Portal, the current legacy picker to select the members will be retired in the following months and the new picker is able to select Security Group. Hence, we need to update this article for now to reduce confusion.