Update Set-TransportRule -SenderAddressLocation section
The documentation is incorrect.
When Header is used during the evaluation of the rule, only property FROM from message header is used. Sender and Reply-To properties are not used.
Confirmation of this can be found in the ADO WI.

It should state:
Header: Only examine senders in the message headers (From field). This is the default value.

The ADO Work Item to correct this mistake can be found here https://o365exchange.visualstudio.com/O365%20Core/_workitems/edit/3510129