Update Set-HostedContentFilterPolicy.md
PG-approved comments on ASF in aka.ms/emailbasics part 2. No, they are not deprecating any ASF- All other articles are upto date except this.. Hence removing the below deprecation NOTE  from 3 segments.. 

"Note: This setting is part of Advanced Spam Filtering (ASF) and will be deprecated. The functionality of this setting will be incorporated into other parts of the filtering stack. We recommend that you leave this setting turned off."