Update Set-MailboxCalendarConfiguration.md
I found this parameter (and others) in this cmdlet last month during my quarterly sweep, but I missed it. 100% this parameter isn't available in Exchange 2016 or 2019, so I removed those Applicable values. The parameter isn't available in the Default parameter set, which affects the Parameter Sets value. Alpha sorting is important in the parameter sets and the parameter sections.